requirements-cuda.txt
onnxruntime-gpu==1.19.2
nvidia-cudnn-cu12==9.1.0.70